Extract ASG ready logic
